+++ ctvbot/proxy.py
@@ -1,16 +1,14 @@
-import random
-from typing import Optional
-
 import logging
+import os
+import random
 
 logger = logging.getLogger(__name__)
 
 
 class ProxyGetter:
-    def __init__(self, pathed_file_name="proxy/proxy_list.txt"):
+    def __init__(self, proxy_file_name="proxy_list.txt"):
         self.proxy_list = []
-        self.pathed_file_name = pathed_file_name
-
+        self.pathed_file_name = os.path.join(os.getcwd(), "proxy", proxy_file_name)
         self.build_proxy_list()
 
     def build_proxy_list(self):
