+++ ctvbot/utils.py
